requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2.7
django-cors-headers==3.10.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.2
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
Pygments==2.10.0
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5